remove postgresv3 package and update the postgres package to support it
