--- packages/spin-postgres3/README.md
@@ -1,3 +1,0 @@
-# Spin Postgres v3
-
-This package provides bindings that enable using the Spin Postgres v3 interface in apps built using the `@spinframework/build-tools`.
