--- packages/spin-postgres3/package.json
@@ -1,33 +1,0 @@
-{
-  "name": "@spinframework/spin-postgres3",
-  "version": "1.0.1",
-  "description": "",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "type": "module",
-  "scripts": {
-    "build": "tsc && cp -r types/* dist/"
-  },
-  "keywords": [],
-  "author": "",
-  "license": " Apache-2.0 WITH LLVM-exception",
-  "devDependencies": {
-    "typescript": "^6.0.3"
-  },
-  "peerDependencies": {
-    "@spinframework/build-tools": ">=1.0.0 <3.0.0"
-  },
-  "files": [
-    "dist",
-    "wit"
-  ],
-  "config": {
-    "witDependencies": [
-      {
-        "witPath": "./wit",
-        "package": "spin:postgres@3.0.0",
-        "world": "spin-postgres3"
-      }
-    ]
-  }
-}
