--- packages/spin-postgres3/tsconfig.json
@@ -1,19 +1,0 @@
-{
-    "compilerOptions": {
-        "rootDir": "./src",
-        "target": "ES2022",
-        "module": "nodenext",
-        "lib": [
-            "ES2022"
-        ],
-        "moduleResolution": "nodenext",
-        "declaration": true,
-        "outDir": "dist",
-        "strict": true,
-        "esModuleInterop": true,
-    },
-    "exclude": [
-        "node_modules",
-        "dist"
-    ]
-}
